Add user guide README and move classes to folder
