+++ build.gradle
@@ -48,7 +48,7 @@ test {
 }
 
 application {
-    mainClassName = "duke.Duke"
+    mainClassName = "duke.main.Launcher"
 }
 
 shadowJar {
