+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
+import duke.duke.DukeException;
+import duke.util.Storage;
+import duke.task.TaskList;
 import duke.task.Task;
 
 /** Represents the command to delete task that inherits from Command. */
