+++ src/main/java/duke/command/PriorityCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
+import duke.duke.DukeException;
+import duke.util.Storage;
+import duke.task.TaskList;
 import duke.task.Task;
 
 public class PriorityCommand extends Command{
@@ -17,12 +17,12 @@ public class PriorityCommand extends Command{
         String message = "";
 
         if (userInput.split(" ").length < 3) {
-            throw new DukeException("OOPS!!! The priority command cannot have parameters. Example command is priority 1 high");
+            throw new DukeException("OOPS!!! The priority command cannot have empty parameters. Example command is priority 1 high");
         }
         String index = userInput.split(" ")[1];
         String priorityString = userInput.split(" ")[2];
         Task task = taskList.updateTaskPriority(Integer.parseInt(index), priorityString);
-        message += "OK, I've marked set the priority of the task to " + task.getPriorityString() + "\n";
+        message += "OK, I've set the priority of the task to " + task.getPriorityString() + "\n";
         message += task.toString() + "\n";
         storage.saveTaskList(taskList);
         return message;
