+++ src/test/java/duke/EventTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.duke.DukeException;
 import duke.task.Event;
 import org.junit.jupiter.api.Test;
