+++ src/test/java/duke/TasklistTest.java
@@ -1,6 +1,7 @@
 package duke;
 
-import duke.task.Task;
+import duke.duke.DukeException;
+import duke.task.TaskList;
 import duke.task.Todo;
 import org.junit.jupiter.api.Test;
